debugged average

+++ week03/dayd03/src/com/skillstorm/arrays/TwoDArrays.java
@@ -198,7 +198,7 @@ public class TwoDArrays {
 			
 		}
 		
-		result = sum / count;
+		result = (double) sum / count;
 		return result;
 	}
